refactor: rename `postPaths.ts` to `getPostPaths.ts`
