+++ src/components/Card.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
+import { getPostUrl } from "@/utils/getPostPaths";
 import { slugifyStr } from "@/utils/slugify";
-import { getPostUrl } from "@/utils/postPaths";
 import Datetime from "./Datetime.astro";
 
 type Props = {
